Update
>Se arreglaron los warnings

+++ giUtilities/src/giVector2i.cpp
@@ -170,20 +170,20 @@ giVector2i::normalize()
 {
   int32 mag = magnitude();
   if (mag > 0) {
-    int32 tmp = 1.0f / mag;
+    int32 tmp = 1 / mag;
     x *= tmp;
     y *= tmp;
   }
   else {
-    x = 0.0f;
-    y = 0.0f;
+    x = 0;
+    y = 0;
   }
 }
 
 int32
 giVector2i::magnitude()
 {
-  return sqrtf(powf(x, 2) + powf(y, 2));
+  return int32(sqrt(pow(x, 2) + pow(y, 2)));
 }
