+++ giUtilities/src/giVector4i.cpp
@@ -236,7 +236,7 @@ giVector4i::crossProd(const giVector4i& inVect)
   return (((y * inVect.z) - (z * inVect.y)),
           ((z * inVect.x) - (x * inVect.z)),
           ((x * inVect.y) - (y * inVect.z)),
-          0.f);
+          0);
 }
 
 void
@@ -244,22 +244,22 @@ giVector4i::normalize()
 {
   int32 mag = magnitude();
   if (mag > 0) {
-    int32 tmp = 1.0f / mag;
+    int32 tmp = 1 / mag;
     x *= tmp;
     y *= tmp;
     z *= tmp;
     w *= tmp;
   }
   else {
-    x = 0.0f;
-    y = 0.0f;
-    z = 0.0f;
-    w = 0.0f;
+    x = 0;
+    y = 0;
+    z = 0;
+    w = 0;
   }
 }
 
 int32
 giVector4i::magnitude()
 {
-  return sqrtf(powf(x, 2) + powf(y, 2) + powf(z, 2) + powf(w, 2));
+  return int32(sqrt(pow(x, 2) + pow(y, 2) + pow(z, 2) + pow(w, 2)));
 }
